Initial commit
This is the first change in getting a simple example project together
for demonstrating various aspects of testing. The main goals of the
project are to demonstrate:

- Shipping a hand-written "client" package for integrating the server,
  including configuration in consuming projects
- Using the `WebApplicationFactory` to create an instance of Kestrel
  running the core server
- Perform closed-loop testing of client code with server code
- Initializing and sharing a database instance across a test collection
- Running acceptance tests, including adding "fault injection" into the
  system via table locking

These are very achievable goals, but we're a long way away from that.
At this time, we have the following:

- A "shared" project for declaring what's shared between the HTTP
  client library and server
- A "core" project that contains the actual implementation
- A "webapi" project for our serverj
- A "producttests" project for unit and integration tests

The project is using EFCore as this is very popular in the community,
in addition to it providing some convenience facilities for managing
the creation and destruction of databases between test runs.