Add query logging
This change adds logging to queries occurring in the `EFTodoService`
class. These are usually helpful when examining logs to understand
what step we're looking at at a glance, rather than having to compare
the test code, implementation code, and log data from tests. This
would also be useful in a production system.