Add generic services fixture
This change allows us to easily use the same configuration that will be
used in the product itself when running our tests. This includes
the registration of all our services in the DI system and usage. This
fixture also registers a custom logging provider to log all messages
that pass through it when:

1. An `Action<string>` is registered on the fixture
2. The `EnableLogging` setting is enabled in `testsettings.json`

Clearly, this also adds the `testsettings.json` file. As part of that
change, we're also eliminating the `TestSettings` class that held
hard-coded information on where to connect, and replacing it with the
`TestSettingsFile` class, which holds a shared copy of the JSON file by
itself (`ServicesFixture` holds more configuration).