Add client code
This change adds a client SDK project in C#, demonstrating how such a
project can be implemented to test the server implementation (from the
perspective of the client) and also wiring up tests against a server.
Some aspects of the client SDK include:

- Working with typed `HttpClient` instances (see below)
- Registering client options
- Calling the server as part of a fully-integrated test pipeline

This included a minor changes to test artifacts, mostly around fixing
resource disposal (that wasn't caught by analyzers for some reason).
That aside, this provides a full client implementation for the (simple)
API and client SDK. The most interesting bits are around options
management in the client.