Add support for Web API integration tests
This change enables integration testing scenarios for the Web API via
the `IntegrationTestFactory` class (thanks to `WebApplicationFactory`).
This allows us to plug into the host creation process, override parts
of the setup so we can configure things like the database connection
string, and replace services as we see fit (such as re-configuring the
`TodoContext` with a new connection string). This class also gives us
an `HttpClient` that is configured to work with the test server.

To make this work, we need to modify our custom logging implementation.
The changes are quite simple: simply re-expose the `ServicesFixture` as
an interface (`ILoggable`), and be sure to implement that interface.

This is a step in the direction of performing end-to-end tests of a
client and server.