Take first steps with Behavior-Driven Development
This change is a foray into Behavior-Driven Development via SpecFlow
and Gherkin. The model isn't particularly complex, so this may not be a
useful example, but is an example nonetheless.

One nice benefit of using SpecFlow is that it has integration with the
xUnit test framework. This is a fairly significant advantage given the
robust features around extensibility and fixtures in xUnit. Combined
with the FluentAssertions library, this produces readable tests with
readable step files. To take advantage of this integration, most of the
"reusable" code from the ProductTests project has been moved into a
project that can be shared across both the Behavior-Driven tests and
the "plain" tests. There were also some warning cleanups.

This provides a simple example of SpecFlow with Gherkin, but there are
additional aspects/features to investigate. This simple toy isn't
necessarily emblematic of what real code would (or could) look like;
it's just here for learning purposes.