Removed the ApiMarker class
The WebApplicationFactory expects to receive a type that identifies the
assembly that contains an ASP.NET application. This was previously
handled by the `ApiMarker` class, but the most idiomatic approach to
this problem is to simply expose the `Program` class in the project
containing the server implementation. This change aligns the tests in
this project with that pattern.